🚀 Version 7.9.0
- AI darkchat

+++ src/events/guild/serverapp/darkchat.js
@@ -16,7 +16,8 @@ module.exports = {
         if (message.channel.id == client.config.DARKCHAT.CHAN) {
 
             if (message.author.bot) return;
-            var msg = message.content;
+            var MessageContent = message.content;
+            var msg_prediction;
 
             var appCount = await appCountModal.findOne({
                 guildID: message.guild.id,
@@ -30,29 +31,26 @@ module.exports = {
                 await appCount.save();
             }
 
-            if (msg.includes('@here') || msg.includes('@everyone')) {
+            //Filter
+            if (MessageContent.includes('@here') || MessageContent.includes('@everyone')) {
 
                 await message.delete();
-                return await message.channel.send({ content: "Cannot mention everyone or here" }).then(msg => {
-                    setTimeout(() => {
-                        msg.delete();
-                    }, 4000);
-                });
+                return await message.channel.send({ content: "Cannot mention everyone or here" })
+                    .then(msg => {
+                        setTimeout(() => {
+                            msg.delete();
+                        }, 4000);
+                    });
             }
 
-            if (msg.length === 3 ) return await message.delete();
+            if (MessageContent.length === 3) return await message.delete();
 
-            if (msg.toLowerCase() === 'hi' || 
-            msg.toLowerCase() === 'hello' || 
-            msg.toLowerCase() === '.',
-            msg.toLowerCase().includes('test')) {
-                await message.delete();
-                return;
-            }
+
+            //LOG
+            await message.delete();
 
             const webhooks = await message.channel.fetchWebhooks();
             const webhook = webhooks.find(wh => wh.token);
-
             if (!webhook) return;
 
             const logChan = await client.channels.cache.get(client.config.DARKCHAT.LOG);
@@ -60,23 +58,44 @@ module.exports = {
             const logEmbed = new EmbedBuilder()
                 .addFields(
                     { name: 'User', value: `<@${message.author.id}>` },
-                    { name: 'Message', value: `${msg}` }
+                    { name: 'Message', value: `${MessageContent}` }
                 )
 
-            await logChan.send({
-                embeds: [logEmbed]
-            });
+            await logChan.send({ embeds: [logEmbed] });
 
-            appCount.darkChatCount += 1;
-            await appCount.save();
+            //API 
+            await fetch(client.mod.DARKCHAT.APILINK, { 
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify({ MessageContent }),
+            }).then((response) => response.json())
+                .then((data) => {
+                    msg_prediction = data.predicted_label;
+                }).catch((error) => {
+                    console.log(`Error: ${error}`);
+                });
 
-            await webhook.send({
-                content: `${msg}`,
-                username: `Anon Msg ${appCount.darkChatCount}`,
-                avatarURL: 'https://thumbs.dreamstime.com/b/illegal-stamp-illegal-round-grunge-stamp-illegal-sign-illegal-136960672.jpg',
-            });
+            if (msg_prediction === "informative") {
 
-            await message.delete();
+                appCount.darkChatCount += 1;
+                await appCount.save();
+
+                await webhook.send({
+                    content: `${MessageContent}`,
+                    username: `Anon Msg ${appCount.darkChatCount}`,
+                    avatarURL: 'https://thumbs.dreamstime.com/b/illegal-stamp-illegal-round-grunge-stamp-illegal-sign-illegal-136960672.jpg',
+                });
+
+            } else {
+                return await message.channel.send({ content: "Content Not Informative!" })
+                    .then(msg => {
+                        setTimeout(() => {
+                            msg.delete();
+                        }, 4000);
+                    });
+            }
         }
     }
 }
